+textcase_LastChar.txt, refactor text casing for logging
